+++ CHANGELOG.md
@@ -15,3 +15,17 @@
 - Changed services logics
 - Removed plugins, system for now
 - Removed shortlinks
+
+## [0.2.3]
+### Added
+- App version checks
+
+### Fixed
+- App tasks module task buffer size limit to 10mb
+
+### Changed
+- Updated app version manifest model kikx_version or min, max
+- Fs service list directory with meta
+
+### MUI Updates
+- Close app on update or uninstalling app